project/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Flask-Cors==4.0.0
waitress==2.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
